Display small farmer recognition document on the producer shop page
I've chosen to display it as part of the header so it's visible in spite
of the tab you are navigating inside the shop because I think this is an
important document.

Besides, putting it on the about page did not seem to provide a nice
user experience.

Finally, if we want to display an image instead of a text indicating
that the producer is a small farmer, we can display it in the right side
of the producer's name. I think that small image placed there would look
nice but we do not have it right now.